+++ game_patch/input/mouse.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void mouse_apply_patch();
+void mouse_get_camera(float& pitch_delta, float& yaw_delta);